Agregando DTOs faltantes y Excepciones

+++ src/main/java/edu/unimagdalena/reservadevuelo/controllers/AerolineaController.java
@@ -1,71 +1,73 @@
 package edu.unimagdalena.reservadevuelo.controllers;
 
 import edu.unimagdalena.reservadevuelo.dto.AerolineaDto;
-import edu.unimagdalena.reservadevuelo.entities.Aerolinea;
-import edu.unimagdalena.reservadevuelo.services.AerolineaService;
+import edu.unimagdalena.reservadevuelo.exeptions.ResourceNotFoundException;
+import edu.unimagdalena.reservadevuelo.services.aerolinea.AerolineaService;
+import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 
 import java.net.URI;
-import java.net.URISyntaxException;
 import java.util.List;
 import java.util.Optional;
 
 @RestController
 @RequestMapping("/api/v1/aerolineas")
+@RequiredArgsConstructor
 public class AerolineaController {
     private final AerolineaService aerolineaService;
 
-    public AerolineaController(AerolineaService aerolineaService) {
-        this.aerolineaService = aerolineaService;
-    }
-
-    @GetMapping()
+    @GetMapping("")
     public ResponseEntity<List<AerolineaDto>> getAllAerolineas() {
         return ResponseEntity.ok(aerolineaService.buscarAerolineas());
     }
 
-    @GetMapping("/{id}")
+    @GetMapping("/id/{id}")
     public ResponseEntity<AerolineaDto> getAerolineaById(@PathVariable Long id) {
         return aerolineaService.buscarAerolineaPorId(id)
-                .map(a -> ResponseEntity.ok().body(a))
-                .orElse(ResponseEntity.notFound().build());
+                .map(aerolinea -> ResponseEntity.ok().body(aerolinea))
+                .orElseThrow(() -> new ResourceNotFoundException("Aerolínea con ID " + id + " no encontrada."));
     }
 
-    @GetMapping("/{name}")
+    @GetMapping("/name/{name}")
     public ResponseEntity<AerolineaDto> getAerolineaByName(@PathVariable String name) {
         return aerolineaService.buscarAerolineasPorNombre(name)
                 .stream().map(a -> ResponseEntity.ok().body(a))
-                .findAny().orElse(ResponseEntity.notFound().build());
+                .findAny().orElseThrow(() -> new ResourceNotFoundException("Aerolínea con nombre " + name + " no encontrada."));
     }
 
-    @PostMapping()
-    public ResponseEntity<AerolineaDto> createAerolinea(@RequestBody AerolineaDto aerolineaDto) throws URISyntaxException {
+    @PostMapping("")
+    public ResponseEntity<AerolineaDto> createAerolinea(@RequestBody AerolineaDto aerolineaDto) {
+        if (aerolineaDto.codigoAerolinea() == null)
+            return ResponseEntity.badRequest().build();
         return createNewAerolinea(aerolineaDto);
     }
 
     public ResponseEntity<AerolineaDto> createNewAerolinea(AerolineaDto aerolineaDto){
-        AerolineaDto newAerolinea = aerolineaService.guardarAerolinea(aerolineaDto);
+        AerolineaDto newAerolineaDto = aerolineaService.guardarAerolinea(aerolineaDto);
         URI location = ServletUriComponentsBuilder.fromCurrentRequest()
                 .path("/{id}")
-                .buildAndExpand(newAerolinea.id())
+                .buildAndExpand(newAerolineaDto.id())
                 .toUri();
-        return ResponseEntity.created(location).body(newAerolinea);
+        return ResponseEntity.created(location).body(newAerolineaDto);
     }
 
-    @PutMapping("/{id}")
+    @PutMapping("/id/{id}")
     public ResponseEntity<AerolineaDto> updateAerolinea(@PathVariable Long id, @RequestBody AerolineaDto aerolineaDto) {
-        Optional<AerolineaDto> updatedAerolinea = aerolineaService.actualizarAerolinea(id, aerolineaDto);
-        return updatedAerolinea
-                .map(a -> ResponseEntity.ok(a))
-                .orElseGet(() -> {
-                    return createNewAerolinea(aerolineaDto);
-                });
+        Optional<AerolineaDto> aerolineaUpdated = aerolineaService.actualizarAerolinea(id, aerolineaDto);
+        return aerolineaUpdated
+                .map(aerolinea -> ResponseEntity.ok()
+                        .body(aerolinea))
+                .orElseThrow(() -> new ResourceNotFoundException("Aerolínea con ID " + id + " no encontrada para actualizar."));
     }
 
-    @DeleteMapping("/{id}")
+    @DeleteMapping("/id/{id}")
     public ResponseEntity<Void> deleteAerolinea(@PathVariable Long id) {
+        boolean exists = aerolineaService.existeAerolineaPorId(id);
+        if (!exists) {
+            throw new ResourceNotFoundException("Aerolínea con ID " + id + " no encontrada para eliminar.");
+        }
         aerolineaService.eliminarAerolinea(id);
         return ResponseEntity.noContent().build();
     }
