+++ src/main/java/edu/unimagdalena/reservadevuelo/controllers/ClienteController.java
@@ -1,9 +1,11 @@
 package edu.unimagdalena.reservadevuelo.controllers;
 
 import edu.unimagdalena.reservadevuelo.dto.ClienteDto;
-import edu.unimagdalena.reservadevuelo.entities.Cliente;
 import edu.unimagdalena.reservadevuelo.exeptions.ClientNotFoundException;
-import edu.unimagdalena.reservadevuelo.services.ClienteService;
+import edu.unimagdalena.reservadevuelo.exeptions.DuplicateResourceException;
+import edu.unimagdalena.reservadevuelo.exeptions.ResourceNotFoundException;
+import edu.unimagdalena.reservadevuelo.services.cliente.ClienteService;
+import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
@@ -15,29 +17,33 @@ import java.util.Optional;
 
 @RequestMapping("/api/v1/clientes")
 @RestController
+@RequiredArgsConstructor
 public class ClienteController {
     private final ClienteService clienteService;
 
-    public ClienteController(ClienteService clienteService) {
-        this.clienteService = clienteService;
-    }
-
-    @GetMapping
+    @GetMapping("")
     public ResponseEntity<List<ClienteDto>> getAllClientes() {
-        return ResponseEntity.ok(clienteService.buscarClientes());
+        List<ClienteDto> clientesDto = clienteService.buscarClientes();
+        if (clientesDto.isEmpty()) {
+            throw new ResourceNotFoundException("No se encontraron clientes registrados.");
+        }
+        return ResponseEntity.ok(clientesDto);
         //return ResponseEntity.status(HttpStatus.OK).body(clienteService.buscarClientes());
     }
 
-    @GetMapping("/{id}")
+    @GetMapping("/id/{id}")
     public ResponseEntity<ClienteDto> getClienteById(@PathVariable Long id) {
         return clienteService.buscarClientePorId(id)
-                .map(c -> ResponseEntity.ok().body(c))
+                .map(clienteDto -> ResponseEntity.ok().body(clienteDto))
                 //.orElseThrow(ClientNotFoundException::new);
                .orElseThrow(() -> new ClientNotFoundException("No se encontro el cliente con el id: " + id));
     }
 
-    @PostMapping()
+    @PostMapping("")
     public ResponseEntity<ClienteDto> createCliente(@RequestBody ClienteDto cliente) throws URISyntaxException {
+        if (clienteService.existeClienteConCorreoElectronico(cliente.correoElectronico())) {
+            throw new DuplicateResourceException("El correo electronico '" + cliente.correoElectronico() + "' ya está registrado.");
+        }
         return createNewCliente(cliente);
         //return ResponseEntity.created(new URI("/api/v1/clientes/" + newCliente.getId())).body(newCliente);
     }
@@ -51,18 +57,20 @@ public class ClienteController {
         return ResponseEntity.created(location).body(newCliente);
     }
 
-    @PutMapping("/{id}")
+    @PutMapping("/id/{id}")
     public ResponseEntity<ClienteDto> updateCliente(@PathVariable Long id, @RequestBody ClienteDto clienteDto) throws URISyntaxException {
         Optional<ClienteDto> clienteUpdate = clienteService.actualizarCliente(id, clienteDto);
         return clienteUpdate
                 .map(c -> ResponseEntity.ok(c))
-                .orElseGet(() -> {
-                    return createNewCliente(clienteDto);
-                });
+                .orElseThrow(() -> new ResourceNotFoundException("Cliente con ID " + id + " no se encontró para actualizar."));
+
     }
 
-    @DeleteMapping("/{id}")
+    @DeleteMapping("/id/{id}")
     public ResponseEntity<Void> deleteCliente(@PathVariable Long id) {
+        if (!clienteService.existeClientePorId(id)) {
+            throw new ResourceNotFoundException("Cliente con ID " + id + " no encontrado para eliminar.");
+        }
         clienteService.eliminarCliente(id);
         return ResponseEntity.noContent().build();
     }
