+++ src/main/java/edu/unimagdalena/reservadevuelo/controllers/ReservaController.java
@@ -1,8 +1,9 @@
 package edu.unimagdalena.reservadevuelo.controllers;
 
 import edu.unimagdalena.reservadevuelo.dto.ReservaDto;
-import edu.unimagdalena.reservadevuelo.entities.Reserva;
-import edu.unimagdalena.reservadevuelo.services.ReservaService;
+import edu.unimagdalena.reservadevuelo.exeptions.ResourceNotFoundException;
+import edu.unimagdalena.reservadevuelo.services.reserva.ReservaService;
+import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
@@ -14,36 +15,45 @@ import java.util.Optional;
 
 @RequestMapping("/api/v1/reserva")
 @RestController
+@RequiredArgsConstructor
 public class ReservaController {
     private final ReservaService reservaService;
 
-    public ReservaController(ReservaService reservaService) {
-        this.reservaService = reservaService;
-    }
-
-    @GetMapping
+    @GetMapping({""})
     public ResponseEntity<List<ReservaDto>> getAllReservas(){
-        return ResponseEntity.ok(reservaService.buscarReservas());
+        List<ReservaDto> reservas = reservaService.buscarReservas();
+        if (reservas.isEmpty()) {
+            throw new ResourceNotFoundException("No se encontraron reservas registradas.");
+        }
+        return ResponseEntity.ok().body(reservas);
     }
 
-    @GetMapping("/{id}")
+    @GetMapping("/id/{id}")
     public ResponseEntity<ReservaDto> getReservaById(@PathVariable Long id){
         return reservaService.buscarReservaPorId(id)
-                .map(r -> ResponseEntity.ok().body(r))
-                .orElse(ResponseEntity.notFound().build());
+                .map(reservaDto -> ResponseEntity.ok().body(reservaDto))
+                .orElseThrow(() -> new ResourceNotFoundException("Reserva con ID " + id + " no encontrada."));
     }
 
-    @GetMapping("/cliente/{clienteId}")
+    @GetMapping("/clienteId/{clienteId}")
     public ResponseEntity<List<ReservaDto>> getReservasByCliente(@PathVariable Long clienteId){
-        return ResponseEntity.ok(reservaService.buscarReservasPorCliente(clienteId));
+        List<ReservaDto> reservas = reservaService.buscarReservasPorCliente(clienteId);
+        if (reservas.isEmpty()) {
+            throw new ResourceNotFoundException("No se encontraron reservas para el cliente con ID " + clienteId);
+        }
+        return ResponseEntity.ok(reservas);
     }
 
-    @GetMapping("/vuelo/{vueloId}")
+    @GetMapping("/vueloId/{vueloId}")
     public ResponseEntity<List<ReservaDto>> getReservasByVuelo(@PathVariable Long vueloId){
-        return ResponseEntity.ok(reservaService.buscarReservasPorVuelo(vueloId));
+        List<ReservaDto> reservas = reservaService.buscarReservasPorVuelo(vueloId);
+        if (reservas.isEmpty()) {
+            throw new ResourceNotFoundException("No se encontraron reservas para el vuelo con ID " + vueloId);
+        }
+        return ResponseEntity.ok(reservas);
     }
 
-    @PostMapping()
+    @PostMapping("")
     public ResponseEntity<ReservaDto> createReserva(@RequestBody ReservaDto reservaDto) throws URISyntaxException {
         return createNewReserva(reservaDto);
     }
@@ -57,18 +67,21 @@ public class ReservaController {
         return ResponseEntity.created(location).body(newReserva);
     }
 
-    @PutMapping("/{id}")
+    @PutMapping("/id/{id}")
     public ResponseEntity<ReservaDto> updateReserva(@PathVariable Long id, @RequestBody ReservaDto reservaDto){
-        Optional<ReservaDto> reservaUpdate = reservaService.buscarReservaPorId(id);
+        Optional<ReservaDto> reservaUpdate = reservaService.actualizarReserva(id, reservaDto);
         return reservaUpdate
-                .map(r -> ResponseEntity.ok(r))
-                .orElseGet(() -> {
-                    return createNewReserva(reservaDto);
-                });
+                .map(ResponseEntity::ok)
+                .orElseThrow(() -> new ResourceNotFoundException("Reserva con ID" + id + " no encontrada para actualizar."));
+
     }
 
-    @DeleteMapping("/{id}")
+    @DeleteMapping("/id/{id}")
     public ResponseEntity<Void> deleteReserva(@PathVariable Long id){
+        if (!reservaService.existeReservaPorId(id)) {
+            throw new ResourceNotFoundException("Reserva con ID " + id + " no encontrada para eliminar.");
+        }
+
         reservaService.eliminarReserva(id);
         return ResponseEntity.noContent().build();
     }
